Move color table things into dedicated cpp/hpp
