+++ src/host/settings.cpp
@@ -5,7 +5,7 @@
 #include "settings.hpp"
 
 #include "..\interactivity\inc\ServiceLocator.hpp"
-#include "../types/inc/utils.hpp"
+#include "../types/inc/colorTable.hpp"
 
 #pragma hdrstop
