+++ src/types/inc/utils.hpp
@@ -45,28 +45,6 @@ namespace Microsoft::Console::Utils
 
     std::string ColorToHexString(const til::color color);
     til::color ColorFromHexString(const std::string_view wstr);
-    bool ColorFromXOrgAppColorName(const std::wstring_view wstr, til::color& color);
-
-    void InitializeCampbellColorTable(const gsl::span<COLORREF> table);
-    void InitializeCampbellColorTableForConhost(const gsl::span<COLORREF> table);
-    void SwapANSIColorOrderForConhost(const gsl::span<COLORREF> table);
-    void Initialize256ColorTable(const gsl::span<COLORREF> table);
-
-    // Function Description:
-    // - Fill the alpha byte of the colors in a given color table with the given value.
-    // Arguments:
-    // - table: a color table
-    // - newAlpha: the new value to use as the alpha for all the entries in that table.
-    // Return Value:
-    // - <none>
-    constexpr void SetColorTableAlpha(const gsl::span<COLORREF> table, const BYTE newAlpha) noexcept
-    {
-        const auto shiftedAlpha = newAlpha << 24;
-        for (auto& color : table)
-        {
-            WI_UpdateFlagsInMask(color, 0xff000000, shiftedAlpha);
-        }
-    }
 
     constexpr uint16_t EndianSwap(uint16_t value)
     {
