+++ src/types/lib/types.vcxproj
@@ -12,6 +12,7 @@
   <ItemGroup>
     <ClCompile Include="..\CodepointWidthDetector.cpp" />
     <ClCompile Include="..\convert.cpp" />
+    <ClCompile Include="..\colorTable.cpp" />
     <ClCompile Include="..\Environment.cpp" />
     <ClCompile Include="..\GlyphWidth.cpp" />
     <ClCompile Include="..\MouseEvent.cpp" />
@@ -39,6 +40,7 @@
     <ClInclude Include="..\IControlAccessibilityInfo.h" />
     <ClInclude Include="..\inc\CodepointWidthDetector.hpp" />
     <ClInclude Include="..\inc\convert.hpp" />
+    <ClInclude Include="..\inc\colorTable.hpp" />
     <ClInclude Include="..\inc\Environment.hpp" />
     <ClInclude Include="..\inc\GlyphWidth.hpp" />
     <ClInclude Include="..\inc\IInputEvent.hpp" />
