+++ src/types/sources.inc
@@ -39,6 +39,7 @@ SOURCES= \
     ..\Viewport.cpp \
     ..\WindowBufferSizeEvent.cpp \
     ..\convert.cpp \
+    ..\colorTable.cpp \
     ..\Utf16Parser.cpp \
     ..\utils.cpp \
     ..\ThemeUtils.cpp \
