+++ src/types/ut_types/UtilsTests.cpp
@@ -6,6 +6,7 @@
 #include "..\..\inc\consoletaeftemplates.hpp"
 
 #include "..\inc\utils.hpp"
+#include "..\inc\colorTable.hpp"
 #include <conattrs.hpp>
 
 using namespace WEX::Common;
